requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
tenacity>=8.2.0
httpx>=0.24.0
aiofiles>=23.0.0
pydantic>=2.0.0